xai-formatting-test/requirements.txt
pydantic-ai
pydantic
python-dotenv
asyncio